Start to refactor dropdown nav item
